Secret type, eslint-disable camelcase multiline
