+++ widgets/secrets/src/types.ts
@@ -1,0 +1,11 @@
+/* eslint-disable camelcase */
+export type Secret = {
+    created_at?: string;
+    created_by?: string;
+    is_hidden_value?: boolean;
+    key?: string;
+    tenant_name?: string;
+    updated_at?: string;
+    visibility?: string;
+};
+/* eslint-enable camelcase */
